Game is completed
The game is done entirely, all that remains is to completely build the game. Tomorrow afternoon I am going to do that and put it as a finished product on my itch.io. I think the project went well, I think it was a great introduction into Unreal and I can now say I have made a game in Unreal engine